First commit to include Blockstack support
Telegram login sessions are saved in user's
Gaia storage, under a json file. User can
log out of Blockstack, and log in to another
browser without having to enter his phone
number again.

Created 3 new Angular services, under 'ng_utils.js',
to enable Blockstack/Gaia support. These are:

* BlockstackAuthManager : Holds Blockstack UserSession,
and responsible for handling login/logout.

* GaiaFileStorage : Implements current interface storage, so it
is analogous to IdbFileStorage, MemoryStorage and
TmpFileStorage. Only available if user is logged
in to Blockstack. Used to get and put files. It
depends on 'BlockstackAuthManager'

* GaiaConfigPersistence: originally webogram would
store session details using the local storage,
using the 'ConfigStorage'. This service was designed
to intercept all config storage calls, to provide
session persistence. It depends on 'GaiaFileStorage'.

Also changed the `AppWelcomeController` to include
a proper welcome page before the telegram login.
This will contain some information about 'blockOgram'
It will also let the user be aware that he is not
logged in to Blockstack yet, and allow him to continue
to the log in process.

Changed color scheme of this fork and some of the icons.